fix: Ignore clicks till the UI is rendered
controller.idle flag is used to synchronize UI interaction and responding
to events. This was controlled from the gameEngine object. This means the
user is allowed to click on cards once the gameEngine has called the gameState
object's functions. By moving this logic to the UI code, we now wait till the
UI has reponded to the event from the gameState object before allowing the user
to continue.